+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1 [unreleased]
 
 - Added `cv` method
+- Added `save_txt` method
 
 ## 0.1.0